Move local strategies to subdirectory
